\name{procSummaryData}
\alias{procSummaryData}
\title{Get Summary CPU data}
\usage{
  procSummaryData(timeSum, timeInterval, cont = F,
    decayParam = 0.5)
}
\arguments{
  \item{timeSum}{the amount of time in seconds to generate
  the plot over}

  \item{timeInterval}{the amount of resolution in the
  plots}

  \item{cont}{logical, whether to start a fresh count or
  pull existing data}
}
\description{
  Get Summary CPU data
}
\author{
  Theodore Van Rooy (\url{http://royaltyanalytics.com/})
}

